+++ methods/src/main/kotlin/methods/angles/AnglesPlans.kt
@@ -18,17 +18,27 @@
 package methods.angles
 
 import engine.expressions.Constants.Pi
+import engine.expressions.Expression
+import engine.expressions.Minus
 import engine.methods.CompositeMethod
 import engine.methods.PublicMethod
 import engine.methods.RunnerMethod
 import engine.methods.plan
+import engine.patterns.AnyPattern
 import engine.patterns.ConstantPattern
 import engine.patterns.FixedPattern
+import engine.patterns.TrigonometricExpressionPattern
 import engine.patterns.degreeOf
 import engine.patterns.withOptionalRationalCoefficient
 import methods.constantexpressions.ConstantExpressionsPlans
+import methods.simplify.SimplifyPlans
 
 enum class AnglesPlans(override val runner: CompositeMethod) : RunnerMethod {
+    /**
+     * Try to convert an angle from degrees to radians.
+     * - Use degree conversion formula
+     * - Simplify the resulting expression
+     */
     @PublicMethod
     ConvertDegreesToRadians(
         plan {
@@ -43,6 +53,11 @@ enum class AnglesPlans(override val runner: CompositeMethod) : RunnerMethod {
         },
     ),
 
+    /**
+     * Try to convert an angle from radians to degrees.
+     * - Use radian conversion formula
+     * - Simplify the resulting expression
+     */
     @PublicMethod
     ConvertRadiansToDegrees(
         plan {
@@ -56,4 +71,68 @@ enum class AnglesPlans(override val runner: CompositeMethod) : RunnerMethod {
             }
         },
     ),
+
+    /**
+     * Try to reduce the argument of a trigonometric expression to a main angle by:
+     * - IF POSSIBLE: reducing it to the first quadrant
+     * - IF POSSIBLE: simplifying the resulting argument
+     * - evaluating the exact value of the resulting main angle
+     */
+    EvaluateExactValuesOfMainAngle(
+        plan {
+            pattern = TrigonometricExpressionPattern(AnyPattern())
+
+            explanation = Explanation.EvaluateExactValueOfMainAngle
+
+            val reductionSteps = engine.methods.stepsproducers.steps {
+                firstOf {
+                    option(AnglesRules.FindReferenceAngleInFirstQuadrantInDegree)
+                    option(AnglesRules.FindReferenceAngleInFirstQuadrantInRadian)
+                }
+                optionally {
+                    applyTo(SimplifyPlans.SimplifyAlgebraicExpression) {
+                        getTrigonometricFunctionFromOptionalNegative(it).firstChild
+                    }
+                }
+                optionally {
+                    applyTo(ConstantExpressionsPlans.SimplifyConstantExpression) {
+                        getTrigonometricFunctionFromOptionalNegative(it).firstChild
+                    }
+                }
+            }
+
+            steps {
+                optionally {
+                    apply(reductionSteps)
+                }
+                applyTo(AnglesRules.EvaluateExactValueOfMainAngle) { getTrigonometricFunctionFromOptionalNegative(it) }
+            }
+        },
+    ),
+
+    /**
+     * Evaluate a trigonometric expression.
+     * - we try to simplify the expression
+     * ┌ IF it can be evaluated as a main angle, we evaluate it exactly.
+     * └ ELSE we try to reduce it to a main angle and then evaluate it.
+     */
+    @PublicMethod
+    EvaluateTrigonometricExpression(
+        plan {
+            pattern = TrigonometricExpressionPattern(AnyPattern())
+
+            explanation = Explanation.EvaluateTrigonometricExpression
+
+            steps {
+                optionally { apply(ConstantExpressionsPlans.SimplifyConstantExpression) }
+
+                shortcut(AnglesRules.EvaluateExactValueOfMainAngle)
+
+                apply(EvaluateExactValuesOfMainAngle)
+            }
+        },
+    ),
 }
+
+private fun getTrigonometricFunctionFromOptionalNegative(expression: Expression) =
+    if (expression is Minus) expression.firstChild else expression