+++ methods/src/main/kotlin/methods/approximation/ApproximationPlans.kt
@@ -21,16 +21,15 @@ import engine.context.Context
 import engine.expressions.DecimalExpression
 import engine.expressions.Expression
 import engine.expressions.RecurringDecimalExpression
+import engine.expressions.containsUnits
 import engine.expressions.xp
 import engine.methods.CompositeMethod
 import engine.methods.PublicMethod
 import engine.methods.RunnerMethod
 import engine.methods.plan
 import engine.methods.stepsproducers.steps
-import engine.patterns.AnyPattern
 import engine.patterns.SignedNumberPattern
 import engine.patterns.condition
-import engine.patterns.degreeOf
 import engine.patterns.productContaining
 import engine.steps.Transformation
 import engine.steps.metadata.Metadata
@@ -129,11 +128,10 @@ enum class ApproximationPlans(override val runner: CompositeMethod) : RunnerMeth
     EvaluateExpressionNumerically(
         object : CompositeMethod(specificPlans = listOf(ConstantExpressionsPlans.SimplifyConstantExpression)) {
             private val numberPattern = SignedNumberPattern()
-            private val degreePattern = degreeOf(AnyPattern())
 
             @Suppress("detekt.ReturnCount")
             override fun run(ctx: Context, sub: Expression): Transformation? {
-                if (numberPattern.matches(ctx, sub) || degreePattern.matches(ctx, sub)) return null
+                if (numberPattern.matches(ctx, sub) || sub.containsUnits()) return null
 
                 val numericValue = sub.doubleValue
                 if (!numericValue.isFinite()) return null