+++ methods/src/test/kotlin/methods/angles/AnglesPlansTest.kt
@@ -64,4 +64,77 @@ class AnglesPlansTest {
             }
         }
     }
+
+    @Test
+    fun testEvaluationOfMainAngles() {
+        testMethod {
+            method = AnglesPlans.EvaluateTrigonometricExpression
+            inputExpr = "sin [10 * /pi/ / 24]"
+
+            check {
+                step { toExpr = "sin [5 /pi/ / 12]" }
+                step { toExpr = "[1 / 4] (sqrt[6] + sqrt[2])" }
+            }
+        }
+
+        testMethod {
+            method = AnglesPlans.EvaluateTrigonometricExpression
+            inputExpr = "cos degree[ 120 ]"
+
+            check {
+                fromExpr = "cos degree[ 120 ]"
+                toExpr = "-[1 / 2]"
+
+                step {
+                    fromExpr = "cos degree[ 120 ]"
+                    toExpr = "-cos (degree[ 180 ] - degree[ 120 ])"
+                }
+
+                step {
+                    fromExpr = "-cos (degree[ 180 ] - degree[ 120 ])"
+                    toExpr = "-cos degree[ 60 ]"
+                }
+
+                step {
+                    fromExpr = "-cos degree[ 60 ]"
+                    toExpr = "-[1 / 2]"
+                }
+            }
+        }
+
+        testMethod {
+            method = AnglesPlans.EvaluateTrigonometricExpression
+            inputExpr = "sin degree[ 157.5 ]"
+
+            check {
+                fromExpr = "sin degree[ 157.5 ]"
+                toExpr = "[1 / 2] sqrt[2 - sqrt[2]]"
+
+                step {
+                    fromExpr = "sin degree[ 157.5 ]"
+                    toExpr = "sin [degree[ 315 ] / 2]"
+                }
+
+                step {
+                    fromExpr = "sin [degree[ 315 ] / 2]"
+                    toExpr = "[1 / 2] sqrt[2 - sqrt[2]]"
+
+                    step {
+                        fromExpr = "sin [degree[ 315 ] / 2]"
+                        toExpr = "sin (degree[ 180 ] - [degree[ 315 ] / 2])"
+                    }
+
+                    step {
+                        fromExpr = "sin (degree[ 180 ] - [degree[ 315 ] / 2])"
+                        toExpr = "sin [degree[ 45 ] / 2]"
+                    }
+
+                    step {
+                        fromExpr = "sin [degree[ 45 ] / 2]"
+                        toExpr = "[1 / 2] sqrt[2 - sqrt[2]]"
+                    }
+                }
+            }
+        }
+    }
 }