+++ CHANGELOG.md
@@ -13,6 +13,7 @@ contents to a new release section (e.g. _[0.x]_).
 
 - Converting angles in degrees to radians. (PLUT-1045)
 - Converting angles in radians to degrees. (PLUT-1046)
+- Evaluate sine, cosine, tangent, and cotangent of main angles. (PLUT-1058)
 - SDK: Add rendering and parsing for angles in degrees (PLUT-1045)
 
 ### Fixed